requirements.txt
# =============================================================================
# Helix — Local Development Dependencies
# These are for working locally: notebooks, testing, linting, type checking.
# Each service has its own requirements inside services/<name>/requirements.txt
# Airflow has its own requirements inside airflow/requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Shared package core (install locally with: pip install -e shared/)
# -----------------------------------------------------------------------------
sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0
psycopg2-binary==2.9.9          # Alembic migrations (sync driver)
minio==7.2.5
pydantic==2.6.3
pydantic-settings==2.2.1
structlog==24.1.0

# -----------------------------------------------------------------------------
# Data + feature engineering
# -----------------------------------------------------------------------------
yfinance==0.2.36
pandas==2.2.1
numpy==1.26.4
scikit-learn>=1.4.0
ta==0.11.0

# -----------------------------------------------------------------------------
# ML
# -----------------------------------------------------------------------------
torch==2.1.0
onnx==1.14.1
onnxruntime==1.16.3
mlflow==2.11.1
boto3==1.34.51

# -----------------------------------------------------------------------------
# API
# -----------------------------------------------------------------------------
fastapi==0.110.0
uvicorn[standard]==0.27.1
slowapi==0.1.9
prometheus-client==0.20.0
httpx==0.27.0

# -----------------------------------------------------------------------------
# Observability
# -----------------------------------------------------------------------------
evidently==0.4.16

# -----------------------------------------------------------------------------
# Notebooks
# -----------------------------------------------------------------------------
jupyterlab==4.1.3
matplotlib==3.8.3
plotly==5.20.0
ipywidgets==8.1.2

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
testcontainers[postgres,minio]==4.4.0

# -----------------------------------------------------------------------------
# Code quality
# -----------------------------------------------------------------------------
ruff==0.3.3
mypy==1.9.0
pre-commit==3.6.2
detect-secrets==1.4.0
types-requests==2.31.0.20240218
pandas-stubs==2.2.1.240316
